# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ResponseMessageWithToken": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nombreCompleto": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"direccion": {
						"type": "string"
					},
					"fotoPerfil": {
						"type": "string"
					},
					"tipoUsuario": {
						"type": "string",
						"enum": [
							"INDIVIDUAL",
							"EMPRESARIAL"
						]
					},
					"nombreEmpresa": {
						"type": "string"
					},
					"nit": {
						"type": "string"
					},
					"rol": {
						"type": "string",
						"enum": [
							"USUARIO",
							"EQUIPO",
							"ADMIN",
							"DESARROLLADOR"
						]
					},
					"rolEquipo": {
						"type": "string",
						"enum": [
							"LECTOR",
							"COMENTARISTA",
							"EDITOR"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"activo",
							"inactivo"
						]
					},
					"empresaId": {
						"type": "number",
						"format": "double"
					},
					"perfilCompleto": {
						"type": "boolean"
					}
				},
				"required": [
					"username",
					"correo",
					"password",
					"nombreCompleto",
					"telefono",
					"direccion",
					"rol"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"currentPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseMessage": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UserDTO_": {
				"properties": {
					"username": {
						"type": "string"
					},
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nombreCompleto": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"direccion": {
						"type": "string"
					},
					"fotoPerfil": {
						"type": "string"
					},
					"tipoUsuario": {
						"type": "string",
						"enum": [
							"INDIVIDUAL",
							"EMPRESARIAL"
						]
					},
					"nombreEmpresa": {
						"type": "string"
					},
					"nit": {
						"type": "string"
					},
					"rol": {
						"type": "string",
						"enum": [
							"USUARIO",
							"EQUIPO",
							"ADMIN",
							"DESARROLLADOR"
						]
					},
					"rolEquipo": {
						"type": "string",
						"enum": [
							"LECTOR",
							"COMENTARISTA",
							"EDITOR"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"activo",
							"inactivo"
						]
					},
					"empresaId": {
						"type": "number",
						"format": "double"
					},
					"perfilCompleto": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
		"/usuarios": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "correo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "nombreCompleto",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "telefono",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "nit",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "estado",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rol",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tipoUsuario",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"INDIVIDUAL",
								"EMPRESARIAL"
							]
						}
					},
					{
						"in": "query",
						"name": "rolEquipo",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"LECTOR",
								"COMENTARISTA",
								"EDITOR"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Usuario creado correctamente",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessageWithToken"
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDTO"
							}
						}
					}
				}
			}
		},
		"/usuarios/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUsuario",
				"responses": {
					"200": {
						"description": "Usuario actualizado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					},
					"404": {
						"description": "Usuario no encontrado"
					},
					"500": {
						"description": "Error del servidor"
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UserDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUsuario",
				"responses": {
					"200": {
						"description": "Usuario eliminado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					},
					"404": {
						"description": "Usuario no encontrado"
					},
					"500": {
						"description": "Error del servidor"
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/usuarios/empresa/{id}": {
			"get": {
				"operationId": "GetEmpresaByIdController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/usuarios/cambiarContrasena": {
			"put": {
				"operationId": "CambiarContrasena",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Cambia la contraseña de un usuario.",
				"tags": [
					"Usuarios",
					"Usuarios"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordDTO"
							}
						}
					}
				}
			}
		},
		"/usuarios/{id}/reactivar": {
			"put": {
				"operationId": "ReactivarUsuario",
				"responses": {
					"200": {
						"description": "Usuario reactivado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					},
					"400": {
						"description": "No se pudo reactivar el usuario"
					},
					"404": {
						"description": "Usuario no encontrado"
					}
				},
				"tags": [
					"Usuarios"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}